// src/app/core/models/mp-nested-member.ts
export interface MpNestedMember {
    id?: number;
    new: boolean;
    email?: string;
    firstName?: string;
    lang?: string;
    lastName?: string;
    fullName?: string;
    mail?: string;
    name?: string;
    mailConfirmation: boolean;
    mobile?: string;
    teamAdmin: boolean;
    teamMembers: any;
}